[rename] dummy_data 타입 폴더 이름 변경 및 mock 폴더로 이동
