+++ src/mock/Friends/type.ts
